Se agrego el poder editar los elementos

+++ script.js
@@ -9,9 +9,10 @@ function listado() {
             <th scope="row">${index}</th>
             <td>${item.titulo}</td>
             <td>${item.autor}</td>
-            <td><button onclick="eliminarLibro(${index})">Eiminar</button></td>
-            <td><button onclick="editarTitulo(${index})">Editar Titulo</button></td>
-            <td><button onclick="editarAutor(${index})">Editar Autor</button></td>
+            <td>
+            <button type="button" class="btn btn-danger" onclick="eliminarLibro(${index})">Eiminar</button>
+            <button type="button" class="btn btn-secondary" data-bs-toggle="modal" data-bs-target="#editModal" onclick="setEditModal(${index})">Editar</button>
+            </td>            
         </tr>`;
     })
 }
@@ -19,6 +20,7 @@ function listado() {
 function agregarLibro() {
     let titulo = document.querySelector('#titulo').value;
     let autor = document.querySelector('#autor').value;
+    let formAgregarNoticia = document.querySelector('#createModal');
 
     libros.push({
         titulo: titulo,
@@ -26,14 +28,40 @@ function agregarLibro() {
     });
 
     localStorage.setItem('libros', JSON.stringify(libros));
-
+    formAgregarNoticia.reset();
     listado();
 }
 
 function eliminarLibro(index) {
-    libros.splice(index, 1);
-    localStorage.setItem('libros', JSON.stringify(libros));
-    listado()
+    let confirmar = confirm("Esta seguro de eliminar?");
+    if (confirmar) {
+        libros.splice(index, 1);
+        localStorage.setItem('libros', JSON.stringify(libros));
+        listado()
+    }
 }
 
 listado();
+
+let editarLibro = function(){
+    let titulo = document.querySelector('#editarTitulo').value;
+    let autor = document.querySelector('#editarAutor').value;
+
+    let index = event.target.dataset.index;
+    libros[index] = {
+        titulo: titulo,
+        autor: autor
+    }
+
+    localStorage.setItem('libros', JSON.stringify(libros));
+    
+    listado(); 
+}
+
+function setEditModal(index){
+    let editBtn = document.getElementById("EditBTN");
+    editBtn.setAttribute("data-index", index);
+
+    document.querySelector('#editarTitulo').value = libros[index].titulo;
+    document.querySelector('#editarAutor').value = libros[index].autor;
+}
